+++ client/src/main/js/config.js
@@ -5,6 +5,7 @@
         [
             'app',
             'ui.router',
+            'ui.carousel',
             'ui.bootstrap',
             'ui.calendar',
             'ui.mask',
@@ -26,7 +27,6 @@
             'toastModule',
             'sidebarModulo',
             'buscaModulo',
-            'aboutModule',
             'footerModule',
             'materialCalendar',
             'calendarioModulo',
@@ -84,7 +84,7 @@
          * State config
          */
         .config(['$stateProvider', '$locationProvider', 'APP_STATES', function ($stateProvider, $locationProvider, APP_STATES) {
-            const view = './view/';
+            const view = 'view/';
 
             $stateProvider
                 .state(APP_STATES.APP.nome, {