+++ CHANGELOG.md
@@ -10,6 +10,11 @@
 * Manutenção: Limpeza da suite de testes do frontend e adicionada cobertura de testes no mesmo.
 * Funcionalidade: Adicionada validação de permissão ao criar e alterar reservas no servidor.
 * Correção: Corrigidos testes no servidor com mudança de Secret do Auth0.
+* Funcionalidade: Adicionada validação para tamanho e formato da imagem de upload.
+* Funcionalidade: Adicionado modal com informações sobre os desenvolvedores e links para github.
+* Correção: Adicionada validação para intervalo em funcionamento de local.
+* Funcionalidade: Adicionada visualização de imagens do local e modal para visualização da imagem 'full size'.
+* Funcionalidade: Adicionada deleção de imagens do local.
 * Funcionalidade: Adicionada verificação da permissão do usuário para criar reserva.
 * Funcionalidade: Adicionada função de fechar o modal da reserva após a criação da mesma.